+++ app/Http/Controllers/Accounts/CashPaymentController.php
@@ -45,7 +45,7 @@ class CashPaymentController extends Controller
 
             $dataSql = Voucher::where('type',self::Constants()['type'])->distinct()->orderby('date','desc');
 
-            $allData = $dataSql->get(['voucher_id','voucher_no','date','remarks']);
+            $allData = $dataSql->get(['voucher_id','voucher_no','date','posted','remarks']);
 
             $recordsTotal = count($allData);
             $recordsFiltered = count($allData);
@@ -60,6 +60,7 @@ class CashPaymentController extends Controller
             }
             $entries = [];
             foreach ($allData as $row) {
+                $posted = $this->getPostedTitle()[$row->posted];
                 $urlEdit = route('accounts.cash-payment.edit',$row->voucher_id);
                 $urlDel = route('accounts.cash-payment.destroy',$row->voucher_id);
 
@@ -80,6 +81,7 @@ class CashPaymentController extends Controller
                 $entries[] = [
                     $row->date,
                     $row->voucher_no,
+                    '<div class="text-center"><span class="badge rounded-pill ' . $posted['class'] . '">' . $posted['title'] . '</span></div>',
                     $row->remarks,
                     $actions,
                 ];
@@ -156,6 +158,7 @@ class CashPaymentController extends Controller
             $voucher_no = self::documentCode(self::Constants()['type'],$max);
             $voucher_id = self::uuid();
 
+            $posted = $request->current_action_id == 'post'?1:0;
             $sr = 1;
             foreach ($request->pd as $pd){
                 $account = ChartOfAccount::where('id',$pd['chart_id'])->first();
@@ -177,6 +180,7 @@ class CashPaymentController extends Controller
                         'company_id' => auth()->user()->company_id,
                         'project_id' => auth()->user()->project_id,
                         'user_id' => auth()->user()->id,
+                        'posted' => $posted,
                     ]);
                     $sr = $sr + 1;
                 }
@@ -226,7 +230,11 @@ class CashPaymentController extends Controller
         }
 
         $data['view'] = false;
-        if(isset($request->view)){
+        $data['posted'] = false;
+        if($data['current']->posted == 1){
+            $data['posted'] = true;
+        }
+        if(isset($request->view) || $data['current']->posted == 1){
             $data['view'] = true;
             $data['permission'] = self::Constants()['view'];
             $data['permission_edit'] = self::Constants()['edit'];
@@ -278,10 +286,14 @@ class CashPaymentController extends Controller
         try {
 
             $firstVoucher = Voucher::where('type',self::Constants()['type'])->where('voucher_id',$id)->first();
+            if($firstVoucher->posted == 1){
+                return $this->jsonErrorResponse($data, 'This voucher have been already posted');
+            }
             $voucher_no = $firstVoucher->voucher_no;
             $voucher_id = $id;
             DB::select("delete FROM `vouchers` where voucher_id = '$voucher_id'");
 
+            $posted = $request->current_action_id == 'post'?1:0;
             $sr = 1;
             foreach ($request->pd as $pd){
                 $account = ChartOfAccount::where('id',$pd['chart_id'])->first();
@@ -303,6 +315,7 @@ class CashPaymentController extends Controller
                         'company_id' => auth()->user()->company_id,
                         'project_id' => auth()->user()->project_id,
                         'user_id' => auth()->user()->id,
+                        'posted' => $posted,
                     ]);
                     $sr = $sr + 1;
                 }
